[IMP] Implementation of dto,dao,sql files, architecture of the proyect including use cases.

+++ microservicio/dominio/src/main/java/com/ceiba/modelo/dto/DtoCompra.java
@@ -11,7 +11,7 @@ public class DtoCompra {
 	
     private Long id;
 	private DtoCliente idCliente;
-    private Long total;
+    private Double total;
 	private LocalDateTime fechaCompra;
 	private LocalDateTime fechaEntrega;
